# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bleibsichtbar Agency API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /portal/content-calendar:
    get:
      operationId: getPortalContentCalendar
      tags: [portal]
      summary: The authenticated customer's real, live Buffer calendar (read-only) — whatever is actually scheduled/published for their channel, not this portal's internal admin drafts
      responses:
        "200":
          description: Buffer calendar
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalContentCalendar"
        "502":
          description: Buffer connection failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /portal/notifications/{id}/read:
    put:
      operationId: markPortalNotificationRead
      tags: [portal]
      summary: Mark one notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /portal/support-tickets:
    get:
      operationId: getPortalSupportTickets
      tags: [portal]
      summary: List the authenticated customer's own support tickets, newest first
      responses:
        "200":
          description: Support tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupportTicket"
    post:
      operationId: createPortalSupportTicket
      tags: [portal]
      summary: Open a new support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupportTicket"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /portal/support-tickets/{id}/messages:
    get:
      operationId: getPortalSupportTicketMessages
      tags: [portal]
      summary: The reply thread for one of the authenticated customer's own tickets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupportTicketMessage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createPortalSupportTicketMessage
      tags: [portal]
      summary: Reply to one of the authenticated customer's own tickets — reopens it if it was resolved/closed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupportTicketMessage"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicketMessage"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /support-tickets/{id}/messages:
    get:
      operationId: getSupportTicketMessages
      tags: [support-tickets]
      summary: The reply thread for one ticket (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupportTicketMessage"
    post:
      operationId: createSupportTicketMessage
      tags: [support-tickets]
      summary: Reply to a ticket — notifies the customer (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupportTicketMessage"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicketMessage"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /portal/instagram:
    get:
      operationId: getPortalInstagram
      tags: [portal]
      summary: Get the authenticated customer's Instagram profile
      parameters:
        - name: days
          in: query
          description: "Relative date-range filter for the Entwicklung/Muster charts — e.g. 5 for \"letzte 5 Tage\". Ignored if since is set. Defaults to the real current calendar month."
          schema:
            type: integer
        - name: since
          in: query
          description: "Format: YYYY-MM-DD — start of an absolute custom date range."
          schema:
            type: string
        - name: until
          in: query
          description: "Format: YYYY-MM-DD — end of an absolute custom date range (defaults to today if since is set but until isn't)."
          schema:
            type: string
      responses:
        "200":
          description: Instagram profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalInstagramProfile"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    PortalContentCalendarPost:
      type: object
      properties:
        id:
          type: string
        date:
          type: string
          description: "Format: YYYY-MM-DD, derived from Buffer's dueAt"
        text:
          type: string
        status:
          type: string
          description: "scheduled | sent"
        thumbnailUrl:
          type: string
          nullable: true
      required:
        - id
        - date
        - text
        - status
        - thumbnailUrl
    PortalContentCalendar:
      type: object
      properties:
        enabled:
          type: boolean
        entries:
          type: array
          items:
            $ref: "#/components/schemas/PortalContentCalendarPost"
      required:
        - enabled
        - entries
...
    SupportTicket:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        category:
          type: string
          description: "invoice | social_media | website | other"
        subject:
          type: string
        message:
          type: string
        status:
          type: string
          description: "open | in_progress | in_review | resolved | closed"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - customerId
        - category
        - subject
        - message
        - status
        - createdAt
        - updatedAt
    SupportTicketMessage:
      type: object
      properties:
        id:
          type: integer
        ticketId:
          type: integer
        senderType:
          type: string
          description: "customer | admin"
        message:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - ticketId
        - senderType
        - message
        - createdAt
    CreateSupportTicketMessage:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    CreateSupportTicket:
      type: object
      properties:
        category:
          type: string
          description: "invoice | social_media | website | other"
        subject:
          type: string
        message:
          type: string
      required:
        - category
        - subject
        - message
...
    PortalInstagramProfile:
      type: object
      properties:
        connected:
          type: boolean
        username:
          type: string
          nullable: true
        profilePictureUrl:
          type: string
          nullable: true
        followers:
          type: integer
          nullable: true
        mediaCount:
          type: integer
          nullable: true
        insights:
          allOf:
            - $ref: '#/components/schemas/PortalInstagramInsights'
          nullable: true
        monthlyHistory:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/PortalMonthlyMetric'
        topMedia:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/PortalInstagramMedia'
        topReels:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/PortalInstagramMedia'
        error:
          type: string
          nullable: true
      required:
        - connected
    PortalInstagramMedia:
      type: object
      properties:
        id:
          type: string
        caption:
          type: string
          nullable: true
        mediaType:
          type: string
        mediaProductType:
          type: string
          nullable: true
        thumbnailUrl:
          type: string
          nullable: true
        permalink:
          type: string
        timestamp:
          type: string
          format: date-time
        likeCount:
          type: integer
        commentsCount:
          type: integer
      required:
        - id
        - mediaType
        - permalink
        - timestamp
        - likeCount
        - commentsCount
    PortalMonthlyMetric:
      type: object
      properties:
        yearMonth:
          type: string
        source:
          type: string
        followers:
          type: integer
          nullable: true
        reach:
          type: integer
          nullable: true
        totalInteractions:
          type: integer
          nullable: true
        profileViews:
          type: integer
          nullable: true
        accountsEngaged:
          type: integer
          nullable: true
        note:
          type: string
          nullable: true
      required:
        - yearMonth
        - source
    PortalInstagramInsights:
      type: object
      properties:
        reach:
          type: integer
          nullable: true
        accountsEngaged:
          type: integer
          nullable: true
        totalInteractions:
          type: integer
          nullable: true
        profileViews:
          type: integer
          nullable: true
        websiteClicks:
          type: integer
          nullable: true
        likes:
          type: integer
          nullable: true
        comments:
          type: integer
          nullable: true
        shares:
          type: integer
          nullable: true
        saves:
          type: integer
          nullable: true
        replies:
          type: integer
          nullable: true
        reachSeries:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/PortalInstagramTimeSeriesPoint'
        followerSeries:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/PortalInstagramTimeSeriesPoint'
        totalInteractionsSeries:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/PortalInstagramTimeSeriesPoint'
        profileViewsSeries:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/PortalInstagramTimeSeriesPoint'
        accountsEngagedSeries:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/PortalInstagramTimeSeriesPoint'
    PortalInstagramTimeSeriesPoint:
      type: object
      properties:
        date:
          type: string
        value:
          type: integer
      required:
        - date
        - value
...